+++ .credo.exs
@@ -1,0 +1,19 @@
+%{
+  configs: [
+    %{
+      name: "default",
+      files: %{
+        included: [
+          "lib/",
+          "test/",
+          "mix.exs"
+        ],
+        excluded: [
+          "_build/",
+          "deps/"
+        ]
+      },
+      strict: true
+    }
+  ]
+}